threads for linux, comfortable texture loading, matrix class
git-svn-id: svn://svn.code.sf.net/p/captainsengine/code/trunk@50 dc34a4ac-3e42-4537-918b-2dea8b01751e

+++ Engine/system/thread.h
@@ -8,7 +8,7 @@ public:
   int create(void (*proc)(void* data), void* data);
   void destroy();
 private:
-  int threadID_;
+  long threadID_;
 };
 
 }
